// src/enums/httpEnum.ts
/**
 * @description: 请求结果集
 */
export enum ResultEnum {
    SUCCESS = 200,
    TIMEOUT = 60000
}

/**
 * @description: 请求方法
 */
export enum RequestHttpEnum {
    GET = 'get',
    POST = 'post',
    PATCH = 'patch',
    PUT = 'put',
    DELETE = 'delete'
}

/**
 * @description:  常用的contentTyp类型
 */
export enum ContentTypeEnum {
    // json
    JSON = 'application/json;charset=UTF-8',
    // text
    TEXT = 'text/plain;charset=UTF-8',
    // xml
    XML = 'application/xml;charset=UTF-8',
    // application/x-www-form-urlencoded 一般配合qs
    FORM_URLENCODED = 'application/x-www-form-urlencoded;charset=UTF-8',
    // form-data  上传
    FORM_DATA = 'multipart/form-data;charset=UTF-8'
}